requirements.txt
# requirements.txt
flask==2.0.1
litellm==1.16.9
python-dotenv==0.19.0